Множественные исправления Множественное исправление кода. Переподготовлен compose. Подготовлено 2а SQL скрипта (src.test.resources"): 1) Создание таблиц. 2) Первоначальное наполнение базы.

+++ src/main/java/com/xandr/pep_aston/dto/UserDto.java
@@ -7,5 +7,5 @@ import lombok.Value;
 @Builder
 public class UserDto {
     String name;
-    Integer pin;
+    Short pin;
 }
