+++ src/main/java/com/xandr/pep_aston/repository/BankAccountRepository.java
@@ -3,9 +3,8 @@ package com.xandr.pep_aston.repository;
 import com.xandr.pep_aston.entity.BankAccount;
 import org.springframework.data.repository.Repository;
 
-public interface BankAccountRepository extends Repository<BankAccount, Integer> {
+public interface BankAccountRepository extends Repository<BankAccount, Long> {
 
     BankAccount save(BankAccount bankAccount);
 
-//    List<BankAccount> findAll();
 }
